requirements.txt
loguru~=0.7.2
markdownify==0.12.1
openai~=1.30.5
openai_whisper
python-dotenv==1.0.1
PyYAML==6.0.1

colorama~=0.4.6
pydub~=0.25.1